The commit will bring the following features :     - An updated version of the underlying database. BDB JE 3.3 is now used.     - Attribute API refactoring providing a better abstraction and offering improved performances.     - A new GUI called the Control-Panel to replace the Status-Panel: the specifications for this       GUI are available on OpenDS Wiki and contains a link to a mockup.       See <https://www.opends.org/wiki/page/ControlPanelUISpecification>.     - Some changes in the replication protocol to implement "Assured Replication Mode". The       specifications are on OpenDS Wiki at <https://www.opends.org/wiki/page/AssuredMode> and section 7       described some of the replication changes required to support this. Assured Replication is not finished,       but the main replication protocol changes to support it are done. As explained by Gilles on an email on       the Dev mailing list (http://markmail.org/message/46rgo3meq3vriy4a), with these changes the newer versions       of OpenDS may not be able to replicate with OpenDS 1.0 instances.     - Support for Service Tags on the platforms where the functionality is available and enabled. Specifications       are published at <https://www.opends.org/wiki/page/OpenDSServiceTagEnabled>. For more information on       Service Tags see <http://wikis.sun.com/display/ServiceTag/Sun+Service+Tag+FAQ>.     - The Admin Connector service. In order to provide agentry of the OpenDS server at any time, a new service       has been added, dedicated to the administration, configuration and monitoring of the server.       An overview of the Admin Connector service and it's use is available on the       OpenDS wiki <https://www.opends.org/wiki/page/ManagingAdministrationTrafficToTheServer>     - Updates to the various command line tools to support the Admin Connector service.     - Some internal re-architecting of the server to put the foundation of future developments such as virtual       directory services. The new NetworkGroups and WorkFlow internal services which have been specified in       <https://www.opends.org/wiki/page/BasicOperationRoutingThroughNetworkGroup> are now implemented.     - Many bug fixes...
This set of changes may have an impact on current clients and deployments, however it should be possible to
upgrade existing 1.0 instances. 

Some of the changes impacting the clients include :
    - Scripts using the command line tools including dsconfig will need to be updated (usually to remove
      some no longer required options).

+++ opendj-sdk/opends/src/server/org/opends/server/backends/jeb/importLDIF/Importer.java
@@ -1097,11 +1097,11 @@ public class Importer implements Thread.UncaughtExceptionHandler {
         long evictPasses = envStats.getNEvictPasses();
         long evictNodes = envStats.getNNodesExplicitlyEvicted();
         long evictBinsStrip = envStats.getNBINsStripped();
-        int cleanerRuns = envStats.getNCleanerRuns();
-        int cleanerDeletions = envStats.getNCleanerDeletions();
-        int cleanerEntriesRead = envStats.getNCleanerEntriesRead();
-        int cleanerINCleaned = envStats.getNINsCleaned();
-        int checkPoints = envStats.getNCheckpoints();
+        long cleanerRuns = envStats.getNCleanerRuns();
+        long cleanerDeletions = envStats.getNCleanerDeletions();
+        long cleanerEntriesRead = envStats.getNCleanerEntriesRead();
+        long cleanerINCleaned = envStats.getNINsCleaned();
+        long checkPoints = envStats.getNCheckpoints();
         if(evictPasses != 0) {
           if(!evicting) {
             evicting=true;
